requirements.txt
discord.py
dotenv
oauth2client
gspread
